hibernate sql 연습
조건1or조건2 검색
조건 is/equals 검색

+++ hibernateTest/src/main/java/com/tistory/herobong/models/user/User.java
@@ -52,7 +52,7 @@ public class User {
 	private String userEmail;
 	
 	@Column(name="user_age", nullable = true, length=3)
-	private Integer userAge;
+	private long userAge;
 	
 	@CreationTimestamp
 	@Column(name="created_at")
